chore: move Neons closer to usage
