+++ src/components/atomic/Background/Background.tsx
@@ -1,21 +1,17 @@
 import React from 'react';
-import { SystemStyleObject, Theme } from '@theme';
 import Box from '@mui/material/Box';
-import Neons from '../Neons/Neons';
+import { Neons } from './Neons/Neons';
 
-export type BackgroundProps = {
-  sx?: SystemStyleObject<Theme>;
-};
-
-const Background: React.FunctionComponent<BackgroundProps> = ({ children, sx }) => (
+const Background: React.FunctionComponent = ({ children }) => (
   <Box
     sx={{
+    display: 'flex',
+    flexDirection: 'column',
       background: 'linear-gradient(152.9deg, #231D40 0%, #151126 52.6%, #201A3A 100%)',
       width: '100vw',
       height: '100vh',
       position: 'fixed',
       boxSizing: 'border-box',
-      ...sx,
     }}
   >
     <Box
